app/requirements.txt
datarobot==3.1.0
langchain==0.0.186
pydantic==1.10.7
streamlit==1.25.0
deep-translator==1.11.4                                                                                                                                                           
plotly==5.16.0
python-dotenv==1.0.0
pandas==1.5.3
numpy==1.23.5
openai==0.27.6
